==> R/5_post_WGCNA_3_find_hubs.R <==
here::i_am("R/5_post_WGCNA_3_find_hubs.R")

########################
# Analyse modules for highly connected genes
# Turquoise: Positive correlation, day + passage
# Anno: Cellular stress response
# Pink: Positive correlation, day only
# Anno: Stemness
# Blue: Negative correlation, day + passage
# Anno: DNA replication
# Red:  Negative correlation, day only
# Anno: Cellular metabolism
########################

# Import packages
library(conflicted)
library(cowplot)
library(data.table)
library(here)
library(openxlsx)
library(tidyverse)
library(WGCNA)

# Load data
gcn <- readRDS(
  file = here::here(
    "output",
    "data_WGCNA",
    "WGCNA_allsamples",
    "GCN_output.RDS"
  )
)

# Set modules
modules_sel <- c(
  "turquoise",
  "pink",
  "blue",
  "red"
)

# Set thresholds
threshold_mm <- 0.8
threshold_gs <- 0.4

# Derive module assignment
genes_sel <- WGCNA::labels2colors(gcn$net$colors) %in% modules_sel
module_assign <- WGCNA::labels2colors(gcn$net$colors)

# Derive module eigengenes
module_eigengenes <- WGCNA::moduleEigengenes(
  gcn$E,
  WGCNA::labels2colors(gcn$net$colors)
) %>%
  .$eigengenes

# Screen for hub genes based on gene significance and module membership
# Within selected modules
# Associate with passage changes

## Calculate intramodular connectivity (degree) of each gene and whole network connectivity
### Intramodular connectivity: abs(cor(GCN$E, use = "pearson")) ^ 6
### kTotal: Whole network connectivity
### kWithin: Within module connectivity
### kOut: Outside module connectivity
### kDiff = kWithin - kOut
## Calculate gene significance of each gene in chosen modules

scores_connectivity <- WGCNA::intramodularConnectivity(
  abs(WGCNA::cor(gcn$E, use = "p"))^6,
  module_assign
) %>%
  cbind(as.numeric(
    WGCNA::cor(
      gcn$targets$Passage,
      gcn$E,
      use = "p"
    )
  )) %>%
  dplyr::rename("gene_significance" = 5) %>%
  cbind(module_assign)

## Calculate module membership for each gene
scores_membership <- WGCNA::signedKME(
  gcn$E,
  module_eigengenes,
  outputColumnName = "membership_"
)

## Combine all data
scores_hub <- cbind(
  scores_connectivity,
  scores_membership
)

scores_hub_sel <- scores_hub[genes_sel, ] %>%
  dplyr::mutate(
    module_assign = module_assign %>%
      factor(levels = modules_sel)
  )

## Plot GS against intramodular connectivity
plot_scores_connectivity <- ggpubr::ggscatter(
  data = scores_hub_sel,
  x = "kWithin",
  y = "gene_significance",
  color = "module_assign",
  add = "reg.line",
  conf.int = TRUE,
  alpha = 0.1,
  size = 1
) +
  ggpubr::stat_cor() +
  ggplot2::scale_color_manual(values = modules_sel) +
  ggplot2::labs(
    x = "Intramodular connectivity",
    y = "Gene significance for passage"
  ) +
  ggplot2::theme(legend.position = "none") +
  ggplot2::facet_wrap(
    ~module_assign,
    ncol = 4,
    nrow = 1,
    labeller = labeller(
      module_assign = c(
        "turquoise" = "Pro-senescence",
        "pink" = "Differentiation potential",
        "blue" = "Proliferative",
        "red" = "Cellular metabolism"
      )
    )
  )

## Plot GS against module membership
plots_module_membership <- purrr::imap(
  list(
    "turquoise" = "membership_turquoise",
    "pink" = "membership_pink",
    "blue" = "membership_blue",
    "red" = "membership_red"
  ),
  \(x, color) {
    ggpubr::ggscatter(
      data = scores_hub_sel,
      x = x,
      y = "gene_significance",
      color = color,
      add = "reg.line",
      conf.int = TRUE,
      alpha = 0.1,
      size = 1
    ) +
      ggplot2::geom_vline(xintercept = threshold_mm) +
      ggplot2::geom_hline(yintercept = c(-1 * threshold_gs, threshold_gs)) +
      ggpubr::stat_cor() +
      ggplot2::labs(
        x = stringr::str_c(
          "Membership in\n",
          color,
          " module"
        ) %>%
          stringr::str_replace_all(c(
            "turquoise" = "pro-senescence",
            "pink" = "differentiation potential",
            "blue" = "proliferative",
            "red" = "cellular metabolism"
          )),
        y = "Gene significance for passage"
      ) +
      ggplot2::theme(legend.position = "none")
  }
)

grid <- cowplot::plot_grid(
  plot_scores_connectivity,
  cowplot::plot_grid(
    plotlist = plots_module_membership,
    ncol = 4,
    nrow = 1
  ),
  ncol = 1,
  nrow = 2,
  labels = "AUTO",
  vjust = 01,
  hjust = -0.5
)

## Screen based on GS and MM metrics in each module
scores_hub_turquoise <- scores_hub %>%
  dplyr::filter(
    abs(gene_significance) > threshold_gs,
    membership_turquoise > threshold_mm,
    module_assign == "turquoise"
  ) %>%
  dplyr::select(gene_significance, membership_turquoise, kWithin) %>%
  dplyr::mutate(entrezid = rownames(.)) %>%
  dplyr::left_join(
    tibble::rownames_to_column(gcn$genes),
    by = dplyr::join_by("entrezid" == "rowname")
  )

scores_hub_pink <- scores_hub %>%
  dplyr::filter(
    abs(gene_significance) > threshold_gs,
    membership_turquoise > threshold_mm,
    module_assign == "pink"
  ) %>%
  dplyr::select(gene_significance, membership_pink, kWithin) %>%
  dplyr::mutate(entrezid = rownames(.)) %>%
  dplyr::left_join(
    tibble::rownames_to_column(gcn$genes),
    by = dplyr::join_by("entrezid" == "rowname")
  )

scores_hub_blue <- scores_hub %>%
  dplyr::filter(
    abs(gene_significance) > threshold_gs,
    membership_turquoise > threshold_mm,
    module_assign == "blue"
  ) %>%
  dplyr::select(gene_significance, membership_blue, kWithin) %>%
  dplyr::mutate(entrezid = rownames(.)) %>%
  dplyr::left_join(
    tibble::rownames_to_column(gcn$genes),
    by = join_by("entrezid" == "rowname")
  )

scores_hub_red <- scores_hub %>%
  dplyr::filter(
    abs(gene_significance) > threshold_gs,
    membership_turquoise > threshold_mm,
    module_assign == "red"
  ) %>%
  dplyr::select(gene_significance, membership_red, kWithin) %>%
  dplyr::mutate(entrezid = rownames(.)) %>%
  dplyr::left_join(
    tibble::rownames_to_column(gcn$genes),
    by = dplyr::join_by("entrezid" == "rowname")
  )

# Export data
openxlsx::write.xlsx(
  list(
    "Turquoise hub candidates" = scores_hub_turquoise,
    "Red hub candidates" = scores_hub_red,
    "Blue hub candidates" = scores_hub_blue,
    "Pink hub candidates" = scores_hub_pink
  ),
  file = here::here(
    "output",
    "data_WGCNA",
    "WGCNA_allsamples",
    "hub_candidates.xlsx"
  )
)

ggsave(
  filename = here::here(
    "output",
    "plots_WGCNA",
    "WGCNA_allsamples",
    "Hub gene screening.png"
  ),
  grid,
  width = 14,
  height = 6.5
)

purrr::imap(
  list(
    "Turquoise hub candidates" = scores_hub_turquoise,
    "Pink hub candidates" = scores_hub_pink,
    "Blue hub candidates" = scores_hub_blue,
    "Red hub candidates" = scores_hub_red
  ),
  \(table, name) {
    data.table::fwrite(
      list(table$symbol),
      file = here::here(
        "output",
        "data_WGCNA",
        "WGCNA_allsamples",
        str_c(name, ".txt")
      ),
      na = ""
    )
  }
)
