==> R/5_post_WGCNA_5_export_Cytoscape.R <==
here::i_am("R/5_post_WGCNA_5_export_Cytoscape.R")

########################
# Export TOMs into edge and node lists for Cytoscape
########################

# Import packages
library(conflicted)
library(here)
library(tidyverse)
library(WGCNA)

# Load data
## GCN with all samples
gcn <- readRDS(
  file = here::here(
    "output",
    "data_WGCNA",
    "WGCNA_allsamples",
    "GCN_output.RDS"
  )
)

# Recalculate TOM
tom <- WGCNA::TOMsimilarityFromExpr(
  datExpr = gcn$E,
  power = 8,
  TOMType = "unsigned",
  verbose = Inf
)

# Export TOMs into edge and node lists for Cytoscape
WGCNA::exportNetworkToCytoscape(
  adjMat = tom,
  edgeFile = here::here(
    "output",
    "data_WGCNA",
    "WGCNA_allsamples",
    "cytoscape_input_edges_module_all.txt"
  ),
  nodeFile = file.path(
    ".",
    "output",
    "data_WGCNA",
    "WGCNA_allsamples",
    "cytoscape_input_nodes_module_all.txt"
  ),
  weighted = TRUE,
  threshold = 0.1,
  # Lower than normal threshold to allow filtering in Cytoscape
  nodeNames = gcn$genes$symbol,
  altNodeNames = rownames(gcn$genes),
  nodeAttr = WGCNA::labels2colors(gcn$net$colors)
)
